print some basic info about Python lists and Python bytes objects

+++ 0x04-python-more_data_structures/103-python.c
@@ -9,8 +9,8 @@ void print_python_bytes(PyObject *p);
  */
 void print_python_list(PyObject *p)
 {
-	Py_ssize_t size, i;
-	/*long int size, i;*/
+	/*Py_ssize_t size, i;*/
+	long int size, i;
 	PyListObject *alloc;
 	PyObject *obj;
 
@@ -36,8 +36,8 @@ void print_python_list(PyObject *p)
  */
 void print_python_bytes(PyObject *p)
 {
-	Py_ssize_t size, i, delim;
-	/*long int size, i, delim;*/
+	/*Py_ssize_t size, i, delim;*/
+	long int size, i, delim;
 	char *str;
 
 	printf("[.] bytes object info\n");
